requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
easyocr==1.7.1
fasttext==0.9.2
filelock==3.12.4
fsspec==2023.9.2
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.5
Jinja2==3.1.2
lazy_loader==0.3
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.26.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.1
pdf2image==1.16.3
Pillow==9.5.0
pybind11==2.11.1
pyclipper==1.3.0.post5
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.22
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.1
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tifffile==2023.9.26
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.0
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.6
wget==3.2
XlsxWriter==3.1.6
